feat(rspack-plugin): enhance package handling and add shared utilities
- Added `@stylexswc/webpack-plugin` as a dependency in the rspack-plugin.
- Introduced `isAllowlistedPackage` function in a new shared module for package validation.
- Updated TypeScript configuration to include the new shared types.
- Modified vitest configuration to resolve shared module paths.
- Refactored the rspack-plugin to utilize the new shared utility for package checks.
- Updated pnpm-lock file to reflect dependency changes.
