+++ packages/rspack-plugin/package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "@stylexjs/babel-plugin": "^0.19.0",
     "@stylexswc/rs-compiler": "0.17.0-rc.3",
+    "@stylexswc/webpack-plugin": "0.17.0-rc.3",
     "loader-utils": "^3.3.1"
   },
   "devDependencies": {
