+++ packages/rspack-plugin/types/webpack-plugin-shared.d.ts
@@ -1,0 +1,3 @@
+declare module '@stylexswc/webpack-plugin/shared' {
+  export function isAllowlistedPackage(resourcePath: string, stylexPackages: string[]): boolean;
+}
