+++ packages/webpack-plugin/package.json
@@ -21,6 +21,19 @@
     "vitest": "^4.1.9",
     "webpack": "^5.108.3"
   },
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
+      "require": "./dist/index.js"
+    },
+    "./shared": {
+      "types": "./dist/shared.d.ts",
+      "import": "./dist/shared.js",
+      "require": "./dist/shared.js"
+    },
+    "./package.json": "./package.json"
+  },
   "files": [
     "dist"
   ],
